+++ output/ArchitectureSmells.csv
@@ -1,0 +1,1 @@
+Project Name,Package Name,Architecture Smell,Cause of the Smell
