+++ output/DesignSmells.csv
@@ -1,0 +1,1 @@
+Project Name,Package Name,Type Name,Design Smell,Cause of the Smell
