+++ src/main/java/com/openkm/cmis/CmisTypeManager.java
@@ -54,7 +54,7 @@ public class CmisTypeManager {
 	public static final String POLICY_TYPE_ID = BaseTypeId.CMIS_POLICY.value();
 	public static final String ITEM_TYPE_ID = BaseTypeId.CMIS_ITEM.value();
 	public static final String SECONDARY_TYPE_ID = BaseTypeId.CMIS_SECONDARY.value();
-	private static final String NAMESPACE = "http://chemistry.apache.org/opencmis/fileshare";
+	public static final String NAMESPACE = "http://chemistry.apache.org/opencmis/fileshare";
 	private Map<String, TypeDefinitionContainerImpl> types;
 	private List<TypeDefinitionContainer> typesList;
 
@@ -75,49 +75,16 @@ public class CmisTypeManager {
 		typeMutability.setCanUpdate(false);
 		typeMutability.setCanDelete(false);
 
+
 		// folder type
-		FolderTypeDefinitionImpl folderType = new FolderTypeDefinitionImpl();
-		folderType.setBaseTypeId(BaseTypeId.CMIS_FOLDER);
-		folderType.setIsControllableAcl(false);
-		folderType.setIsControllablePolicy(false);
-		folderType.setIsCreatable(true);
-		folderType.setDescription("Folder");
-		folderType.setDisplayName("Folder");
-		folderType.setIsFileable(true);
-		folderType.setIsFulltextIndexed(false);
-		folderType.setIsIncludedInSupertypeQuery(true);
-		folderType.setLocalName("Folder");
-		folderType.setLocalNamespace(NAMESPACE);
-		folderType.setIsQueryable(false);
-		folderType.setQueryName("cmis:folder");
-		folderType.setId(FOLDER_TYPE_ID);
-		folderType.setTypeMutability(typeMutability);
+		FolderTypeDefinitionImpl folderType = DefinitionImpl.createFolderDefinition(typeMutability);
 
 		addBasePropertyDefinitions(folderType);
 		addFolderPropertyDefinitions(folderType);
 
 		addTypeInteral(folderType);
-
 		// document type
-		DocumentTypeDefinitionImpl documentType = new DocumentTypeDefinitionImpl();
-		documentType.setBaseTypeId(BaseTypeId.CMIS_DOCUMENT);
-		documentType.setIsControllableAcl(false);
-		documentType.setIsControllablePolicy(false);
-		documentType.setIsCreatable(true);
-		documentType.setDescription("Document");
-		documentType.setDisplayName("Document");
-		documentType.setIsFileable(true);
-		documentType.setIsFulltextIndexed(false);
-		documentType.setIsIncludedInSupertypeQuery(true);
-		documentType.setLocalName("Document");
-		documentType.setLocalNamespace(NAMESPACE);
-		documentType.setIsQueryable(false);
-		documentType.setQueryName("cmis:document");
-		documentType.setId(DOCUMENT_TYPE_ID);
-		documentType.setTypeMutability(typeMutability);
-
-		documentType.setIsVersionable(false);
-		documentType.setContentStreamAllowed(ContentStreamAllowed.ALLOWED);
+		DocumentTypeDefinitionImpl documentType = DefinitionImpl.createDocumentDefinition(typeMutability);
 
 		addBasePropertyDefinitions(documentType);
 		addDocumentPropertyDefinitions(documentType);
@@ -125,21 +92,7 @@ public class CmisTypeManager {
 		addTypeInteral(documentType);
 
 		// relationship types
-		RelationshipTypeDefinitionImpl relationshipType = new RelationshipTypeDefinitionImpl();
-		relationshipType.setBaseTypeId(BaseTypeId.CMIS_RELATIONSHIP);
-		relationshipType.setIsControllableAcl(false);
-		relationshipType.setIsControllablePolicy(false);
-		relationshipType.setIsCreatable(false);
-		relationshipType.setDescription("Relationship");
-		relationshipType.setDisplayName("Relationship");
-		relationshipType.setIsFileable(false);
-		relationshipType.setIsIncludedInSupertypeQuery(true);
-		relationshipType.setLocalName("Relationship");
-		relationshipType.setLocalNamespace(NAMESPACE);
-		relationshipType.setIsQueryable(false);
-		relationshipType.setQueryName("cmis:relationship");
-		relationshipType.setId(RELATIONSHIP_TYPE_ID);
-		relationshipType.setTypeMutability(typeMutability);
+		RelationshipTypeDefinitionImpl relationshipType = DefinitionImpl.createRelationDefinition(typeMutability);
 
 		addBasePropertyDefinitions(relationshipType);
 
@@ -147,21 +100,7 @@ public class CmisTypeManager {
 		// addTypeInteral(relationshipType);
 
 		// policy type
-		PolicyTypeDefinitionImpl policyType = new PolicyTypeDefinitionImpl();
-		policyType.setBaseTypeId(BaseTypeId.CMIS_POLICY);
-		policyType.setIsControllableAcl(false);
-		policyType.setIsControllablePolicy(false);
-		policyType.setIsCreatable(false);
-		policyType.setDescription("Policy");
-		policyType.setDisplayName("Policy");
-		policyType.setIsFileable(false);
-		policyType.setIsIncludedInSupertypeQuery(true);
-		policyType.setLocalName("Policy");
-		policyType.setLocalNamespace(NAMESPACE);
-		policyType.setIsQueryable(false);
-		policyType.setQueryName("cmis:policy");
-		policyType.setId(POLICY_TYPE_ID);
-		policyType.setTypeMutability(typeMutability);
+		PolicyTypeDefinitionImpl policyType = DefinitionImpl.createPolicyDefinition(typeMutability);
 
 		addBasePropertyDefinitions(policyType);
 
@@ -169,21 +108,8 @@ public class CmisTypeManager {
 		// addTypeInteral(policyType);
 
 		// item type
-		ItemTypeDefinitionImpl itemType = new ItemTypeDefinitionImpl();
-		itemType.setBaseTypeId(BaseTypeId.CMIS_ITEM);
-		itemType.setIsControllableAcl(false);
-		itemType.setIsControllablePolicy(false);
-		itemType.setIsCreatable(true);
-		itemType.setDescription("Item");
-		itemType.setDisplayName("Item");
-		itemType.setIsFileable(true);
-		itemType.setIsIncludedInSupertypeQuery(true);
-		itemType.setLocalName("Item");
-		itemType.setLocalNamespace(NAMESPACE);
-		itemType.setIsQueryable(false);
-		itemType.setQueryName("cmis:item");
-		itemType.setId(ITEM_TYPE_ID);
-		itemType.setTypeMutability(typeMutability);
+
+		ItemTypeDefinitionImpl itemType = DefinitionImpl.createItemDefinition(typeMutability);
 
 		addBasePropertyDefinitions(itemType);
 
@@ -191,21 +117,7 @@ public class CmisTypeManager {
 		// addTypeInteral(itemType);
 
 		// secondary type
-		SecondaryTypeDefinitionImpl secondaryType = new SecondaryTypeDefinitionImpl();
-		secondaryType.setBaseTypeId(BaseTypeId.CMIS_ITEM);
-		secondaryType.setIsControllableAcl(false);
-		secondaryType.setIsControllablePolicy(false);
-		secondaryType.setIsCreatable(true);
-		secondaryType.setDescription("Secondary");
-		secondaryType.setDisplayName("Secondary");
-		secondaryType.setIsFileable(false);
-		secondaryType.setIsIncludedInSupertypeQuery(true);
-		secondaryType.setLocalName("Secondary");
-		secondaryType.setLocalNamespace(NAMESPACE);
-		secondaryType.setIsQueryable(false);
-		secondaryType.setQueryName("cmis:secondary");
-		secondaryType.setId(SECONDARY_TYPE_ID);
-		secondaryType.setTypeMutability(typeMutability);
+		SecondaryTypeDefinitionImpl secondaryType = DefinitionImpl.createSecondaryDefinition(typeMutability);
 
 		addBasePropertyDefinitions(secondaryType);
