+++ src/main/java/com/openkm/extension/frontend/client/widget/messaging/userinfo/ProposedQueryReceivedUserInfoValue.java
@@ -32,7 +32,7 @@ import com.openkm.frontend.client.extension.widget.userinfo.UserInfoExtension;
  */
 public class ProposedQueryReceivedUserInfoValue extends UserInfoExtension {
 
-	private HTML htmlValue;
+
 
 	/**
 	 * MessageUserInfoValue
