+++ src/main/java/com/openkm/frontend/client/widget/massive/CategoriesPopup.java
