+++ src/test/resources/templates/sample_out.html
@@ -8,7 +8,7 @@
       </tr>
       <tr>
         <td><b>Bird Date</b></td>
-        <td>26.09.2020</td>
+        <td>25-Mar-2022</td>
       </tr>
       <tr>
         <td><b>Language</b></td>
